+++ dev/src/clasew/examples5.clj
@@ -1,9 +1,9 @@
 (ns
   ^{:author "Frank V. Castellucci"
-      :doc "clasew example 4 - An Excel Example"}
-  clasew.examples4
+      :doc "clasew example 5 - An Apple Numbers Example"}
+  clasew.examples5
   (:require [clasew.spreads :as cs]
-            [clasew.excel :as es]
+            [clasew.numbers :as ans]
             [clojure.pprint :refer :all])
   )
 
@@ -16,48 +16,48 @@
 ;; ---------------- BASIC CALLING ---------------------------------------------
 
 
-(def sample1 (cs/clasew-script "clasew-ex4-sample1.xlsx"
+(def sample1 (cs/clasew-script "clasew-ex5-sample1.numbers"
       cs/create cs/no-open "path to desktop"
       {:handler_list [] :arg_list []}))
 
-; (p sample1)
-; (p (es/clasew-excel-call! sample1))
+;(p sample1)
+;(p (ans/clasew-numbers-call! sample1))
 
 ;; Handler Chaining - Create, put values, get first row, save, quit
 
 (def datum (into []
       (map vec
-           (for [i (range 10)]
-             (repeatedly 10 #(rand-int 100))))))
+           (for [i (range 5)]
+             (repeatedly 5 #(rand-int 100))))))
 
-;(p datum)
+; (p datum)
 
-(def sample2 (cs/clasew-script "clasew-ex4-sample2.xlsx"
+(def sample2 (cs/clasew-script "clasew-ex5-sample2.numbers"
       cs/create cs/no-open "path to desktop"
-      (cs/clasew-handler [[:put-range-data "Sheet1" "A1:J10" datum]
-                             [:get-range-data "Sheet1" "A1:A10"]
+      (cs/clasew-handler [[:put-range-data "Sheet 1" "A1:J10" datum]
+                             [:get-range-data "Sheet 1" "A1:A10"]
                              [:save-quit]])))
 
 ;(p sample2)
-;(p (es/clasew-excel-call! sample2))
+;(p (ans/clasew-numbers-call! sample2))
 
-(def sample2a (cs/clasew-script "clasew-ex4-sample2.xlsx"
+(def sample2a (cs/clasew-script "clasew-ex5-sample2.numbers"
       cs/no-create cs/open "path to desktop"
-      (cs/clasew-handler [[:get-range-formulas "Sheet1" "used"]
+      (cs/clasew-handler [[:get-range-formulas "Sheet 1" "used"]
                                 [:quit-no-save]])))
 
 
 ;(p sample2a)
-;(p (es/clasew-excel-call! sample2a))
+;(p (ans/clasew-numbers-call! sample2a))
 ;; Demo - Open, Read, Quit
 
-(def sample3 (cs/clasew-script "clasew-ex4-sample2.xlsx"
+(def sample3 (cs/clasew-script "clasew-ex5-sample2.numbers"
       cs/no-create cs/open "path to desktop"
       (cs/clasew-handler [[:all-book-info]
                              [:quit]])))
 
 ; (p sample3)
-; (p (es/clasew-excel-call! sample3))
+; (p (ans/clasew-numbers-call! sample3))
 
 ;; Demo pushing your own agenda
 
@@ -67,26 +67,30 @@
     return {my_script: \"No you don't\"}
   end run")
 
-(def sample4 (cs/clasew-script "clasew-ex4-sample2.xlsx"
+(def sample4 (cs/clasew-script "clasew-ex5-sample2.numbers"
       cs/no-create cs/open "path to desktop"
       (cs/clasew-handler [[:run-script my_script []]
                                 [:quit]])))
 
-; (p (es/clasew-excel-call! sample4))
+; (p (ans/clasew-numbers-call! sample4))
 
 ;; Demo HOF - Create, put values, get first row, save, quit
 
-(def wrkbk-name "clasew-ex4-sample5.xlsx")
+(def wrkbk-name "clasew-ex5-sample5.numbers")
 (def wrkbk-path "path to desktop")
 
+
 (def sample5 (cs/create-wkbk wrkbk-name wrkbk-path
-                     (cs/chain-put-range-data "Sheet1" datum)
+                     (cs/chain-put-range-data "Sheet 1" datum)
                      [:save-quit]))
 
 
-; (def s5r (es/clasew-excel-call! sample5))
+(p sample5)
 
-; (p (cs/chain-put-range-data "Sheet1" datum))
+(p (ans/clasew-numbers-call! sample5))
+; (def s5r (ans/clasew-numbers-call! sample5))
+
+; (p (cs/chain-put-range-data "Sheet 1" datum))
 
 ;; Demo misc forms
 
@@ -129,8 +133,8 @@
 
 ;; Put it together
 
-(def sample5a (cs/create-wkbk "clasew-ex4-sample5a.xlsx" wrkbk-path
-      (cs/chain-put-range-data "Sheet1"
+(def sample5a (cs/create-wkbk "clasew-ex5-sample5a.numbers" wrkbk-path
+      (cs/chain-put-range-data "Sheet 1"
                                (cs/extend-columns
                                 (cs/extend-rows
                                  datum 0 0 cs/sum-by-row cs/avg-by-row)
@@ -138,7 +142,7 @@
                                [:save-quit]))
 
 #_(p sample5a)
-#_(p (es/clasew-excel-call! sample5a))
+#_(p (ans/clasew-numbers-call! sample5a))
 
 ;; Demo - cleaning up results
 
@@ -149,27 +153,26 @@
 (def sample6 (cs/open-wkbk wrkbk-name wrkbk-path
                      [:all-book-info] [:quit]))
 
-;(def s6r (cs/clean-result (es/clasew-excel-call! sample6)))
+;(def s6r (cs/clean-result (ans/clasew-numbers-call! sample6)))
 
 ;(p s6r)
 
 (def sample7 (cs/create-wkbk wrkbk-name wrkbk-path
                            (cs/chain-add-sheet
-                            "Before Sheet1" :before "Sheet1"
-                            "After Before Sheet1" :after "Before Sheet1"
+                            "After Sheet 1" :after "Sheet 1"
+                            "Before After Sheet 1" :before "After Sheet 1"
                             "The End" :at :end
                             "The Beginning" :at :beginning
-                            "Also Before Sheet1" :at 4)
+                            "Towards last" :at 4)
                              [:save-quit]))
 
 
 ; (p sample7)
 
-; (p (cs/clean-result (es/clasew-excel-call! sample7)))
+; (p (cs/clean-result (ans/clasew-numbers-call! sample7)))
 
 (def sample8 (cs/open-wkbk wrkbk-name wrkbk-path
-                 (cs/chain-delete-sheet "Sheet1")))
+                 (cs/chain-delete-sheet "Sheet 1")))
 
-; (p sample8)
 
-; (p (cs/clean-result (es/clasew-excel-call! sample8)))
+; (p (cs/clean-result (ans/clasew-numbers-call! sample8)))